Add Content Security Policy and vendor some resources
Adding a Content Security Policy may help when the running code somehow
manages to escape the sandbox and it also should help in the future for
protecting a bit against malicious SVGs - maybe, we should probably use
a sandboxed iframe for that purpose. In any case, it is a good idea as a
defense in depth strategy.

There is a rule "style-src 'unsafe-inline'". This is intentional:
otherwise we can't set custom CSS properties in the SVG. It shouldn't
result in any serious issues because none of the other CSP policies
allow arbitrary content to be loaded. For details, see:
https://stackoverflow.com/questions/30653698/csp-style-src-unsafe-inline-is-it-worth-it

I've vendored some of the external dependencies (Bootstrap, jQuery, etc)
to make the CSP simpler and (maybe) to improve load time. The browser
already has a connection to the origin server, so can quickly download
these dependencies over the same connection.

TODO: remove some of these dependencies. I don't think we really _need_
all the external JS we're including now.