Move runner into separate worker thread
This has a few benefits (like better parallelism) but most importantly
it allows the runner to block. Which in turn makes it possible to
directly execute WASI in a browser (`time.Sleep` would be impossible to
implement otherwise).

There is one major downside though: it relies on SharedArrayBuffer which
is not supported in VSCode at the moment. So this change cannot
currently work in the TinyGo extension. For more information, see this
discussion:
https://github.com/microsoft/vscode-discussions/discussions/156