Implement zoom/pan
This uses the same kind of interface as EasyEDA and KiCad: scrolling
zooms at the location of the mouse cursor, and dragging with the right
mouse button pans.

Also remove the restriction that parts can't be moved outside the view.
It's now much easier to fix that if it happens unintentionally, and it
matches EDA tools.