Add power consumption estimates
Add a 'Power' tab, that lists current consumption for most built-in
parts. The goal (among others) is to make it easier to estimate power
consumption for things like LED strips, which can consume a ton of
power. It may also be helpful to know how much power is consumed by a
display backlight vs the rest of the system, for example.

I got these numbers mostly by measuring them directly, and sometimes by
reading datasheets.

Right now energy consumption is a separate tab, but it might actually be
better to merge it into the properties pane. After all, energy
consumption is also a property of a device.