Use array instead of map in state.parts to keep part order
This changes the format of the simulator state, changing from a map {}
to an array []. To keep backwards compatibility, old state is
automatically upgraded to the new format.

This is needed because while JavaScript keeps map keys sorted, JSON does
not. So after saving and reloading state (or sharing!) the order of
parts would change.

I hope I didn't break anything with this change...