Add some server-side tracking
Store how often a compilation request is done for a particular
page/board/etc and whether it was as a result of modifying the sample.

I've tried to make this as privacy-preserving as reasonably possible.
Timestamps are rounded to a day, and no source code is stored (only
whether it was modified).

There is a very bare bones chart on the /track URL, which should get
expanded as we have more data and can actually show interesting charts.